wip on solving the chaning of two thunks one for post and one for delete

+++ src/app/api/sessions/route.ts
@@ -5,10 +5,7 @@ import { NextResponse, NextRequest } from 'next/server';
 import _ from 'lodash';
 import { ObjectId } from 'bson';
 import { session } from '@/app/_types/types';
-import {
-  createMissingObjIds,
-  prepareTemplate,
-} from '@/app/api/_shared';
+import { createMissingObjIds, prepareTemplate } from '@/app/api/_shared';
 export async function GET(req: NextRequest) {
   await dbConnect();
   try {
@@ -46,7 +43,7 @@ export async function POST(req: NextRequest) {
           message: 'Request data should be an array',
         }),
         {
-          status: 500,
+          status: 400,
           headers: {
             'Content-Type': 'application/json',
           },
@@ -64,6 +61,7 @@ export async function POST(req: NextRequest) {
 
     let tempDoc;
     if (!existingTemplate) {
+      // complete this function, should take title, description from user input
       const template = prepareTemplate(templateId, 'tets', 'description', data);
       tempDoc = await Template.create(template);
     }
@@ -88,6 +86,7 @@ export async function POST(req: NextRequest) {
         message: 'data added',
         template: tempDoc || updatedDoc,
         new_sessions: newData,
+        all_data: data,
       }),
       {
         status: 200,
@@ -113,11 +112,44 @@ export async function DELETE(req: NextRequest) {
   await dbConnect();
 
   try {
-    const { _id } = await req.json();
-    const sessions = await Session.findOneAndDelete({ _id });
-    if (!sessions) throw new Error('id given was not found');
+    const url = new URL(req.url);
+    const params = url.searchParams;
+    const _id = params.get('_id');
+    if (!_id) {
+      return NextResponse.json(
+        { error: "Missing '_id' parameter" },
+        { status: 400 }
+      );
+    }
 
-    return NextResponse.json(sessions);
+    // const { _id } = await req.json();
+    console.log('id', _id);
+    const sessions = await Session.findOneAndDelete({ _id });
+    if (!sessions) {
+      return new NextResponse(
+        JSON.stringify({
+          message: `${_id} was not found in the database`,
+        }),
+        {
+          status: 404,
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        }
+      );
+    }
+    return new NextResponse(
+      JSON.stringify({
+        message: `Object Id: ${_id} was removed from the database`,
+        _id,
+      }),
+      {
+        status: 200,
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      }
+    );
   } catch (err) {
     if (err instanceof Error) {
       return NextResponse.json({ error: err.message });
